refactor: mutualize code for panel field source and table row field source
