+++ src/main/java/fr/siamois/ui/form/fieldsource/TableRowFieldSource.java
@@ -1,45 +1,22 @@
 package fr.siamois.ui.form.fieldsource;
 
 import fr.siamois.domain.models.form.customfield.CustomField;
-import fr.siamois.domain.models.form.customform.DependsOnJson;
-import fr.siamois.domain.models.form.customform.EnabledWhenJson;
 import fr.siamois.ui.form.dto.CustomColUiDto;
-import fr.siamois.ui.form.dto.CustomFormPanelUiDto;
-import fr.siamois.ui.form.dto.CustomRowUiDto;
 import fr.siamois.ui.form.dto.FormUiDto;
 import fr.siamois.ui.table.TableDefinition;
 import fr.siamois.ui.table.column.FormFieldColumn;
 import fr.siamois.ui.table.column.TableColumn;
-
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.Map;
 
 /**
  * FieldSource pour une ligne de tableau.
  *
  * Il se base sur :
  *  - la définition globale de la table (colonnes → CustomField)
- *  - optionnellement, un CustomForm spécifique à la ligne pour récupérer les EnabledWhenJson.
+ *  - optionnellement, un CustomForm spécifique à la ligne pour récupérer les EnabledWhenJson/DependsOnJson.
  */
-public class TableRowFieldSource implements FieldSource {
+public class TableRowFieldSource extends AbstractFieldSource {
 
     private final TableDefinition tableDefinition;
-
-    /**
-     * Map par id de field pour résolution rapide.
-     */
-    private final Map<Long, CustomField> byId = new HashMap<>();
-
-    /**
-     * Map des règles enabledWhen par field (optionnel).
-     */
-    private final Map<CustomField, EnabledWhenJson> enabledByField = new HashMap<>();
-
-    /**
-     * Map des règles dependsOn par field (optionnel).
-     */
-    private final Map<CustomField, DependsOnJson> dependsOnByField = new HashMap<>();
 
     public TableRowFieldSource(TableDefinition tableDefinition) {
         this(tableDefinition, null);
@@ -47,13 +24,13 @@ public class TableRowFieldSource implements FieldSource {
 
     /**
      * @param tableDefinition définition globale des colonnes
-     * @param rowSpecificForm CustomForm spécifique à la ligne (si tu veux récupérer les EnabledWhenJson depuis le form)
+     * @param rowSpecificForm CustomForm spécifique à la ligne (si tu veux récupérer les EnabledWhenJson/DependsOnJson depuis le form)
      */
     public TableRowFieldSource(TableDefinition tableDefinition, FormUiDto rowSpecificForm) {
         this.tableDefinition = tableDefinition != null ? tableDefinition : new TableDefinition();
         indexFromTableDefinition();
         if (rowSpecificForm != null) {
-            indexEnabledWhenFromForm(rowSpecificForm);
+            walkForm(rowSpecificForm, this::indexSpecsForKnownField);
         }
     }
 
@@ -68,69 +45,10 @@ public class TableRowFieldSource implements FieldSource {
         }
     }
 
-
-    private void indexEnabledWhenFromForm(FormUiDto form) {
-        if (form.getLayout() == null) {
+    private void indexSpecsForKnownField(CustomField field, CustomColUiDto column) {
+        if (field.getId() == null || !byId.containsKey(field.getId())) {
             return;
         }
-
-        for (CustomFormPanelUiDto panel : form.getLayout()) {
-            processPanel(panel);
-        }
-    }
-
-    private void processPanel(CustomFormPanelUiDto panel) {
-        if (panel.getRows() == null) {
-            return;
-        }
-
-        for (CustomRowUiDto row : panel.getRows()) {
-            processRow(row);
-        }
-    }
-
-    private void processRow(CustomRowUiDto row) {
-        if (row.getColumns() == null) {
-            return;
-        }
-
-        for (CustomColUiDto column : row.getColumns()) {
-            processColumn(column);
-        }
-    }
-
-    private void processColumn(CustomColUiDto column) {
-        CustomField field = column.getField();
-        if (field == null || field.getId() == null || !byId.containsKey(field.getId())) {
-            return;
-        }
-
-        if (column.getEnabledWhenSpec() != null) {
-            enabledByField.put(field, column.getEnabledWhenSpec());
-        }
-        if (column.getDependsOnSpec() != null) {
-            dependsOnByField.put(field, column.getDependsOnSpec());
-        }
-    }
-
-
-    @Override
-    public Collection<CustomField> getAllFields() {
-        return byId.values();
-    }
-
-    @Override
-    public CustomField findFieldById(Long id) {
-        return byId.get(id);
-    }
-
-    @Override
-    public EnabledWhenJson getEnabledSpec(CustomField field) {
-        return enabledByField.get(field);
-    }
-
-    @Override
-    public DependsOnJson getDependsOnSpec(CustomField field) {
-        return dependsOnByField.get(field);
+        registerSpecs(field, column);
     }
 }
